+++ domain/src/main/java/com/emil/domain/model/PostResponse.kt
@@ -3,6 +3,6 @@ package com.emil.domain.model
 data class PostResponse (val postId:Long,val authorUsername:String,val authorAvatarUrl:String,
                          val imageUrl:String?,val videoUrl:String?,val audioUrl:String?,val voiceUrl:String?,
                          val text:String?,val publishDate:String,
-                         val likesCount:Int,val repostsCount:Int)
+                         val likesCount:Long,val repostsCount:Long)
